app/requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
plotly>=5.18.0
matplotlib>=3.8.0
numpy>=1.26.0
requests>=2.31.0
urllib3>=2.0.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
